refactor: externaize UtxoIcon component and utxoTags function
